Create Build a Pokémon Search App
Build a Pokémon Search App
In this project, you'll build an app that will search for Pokémon by name or ID and display the results to the user. To retrieve the Pokémon data and images, you'll use freeCodeCamp's PokéAPI Proxy.

Note: The first 13 steps must be completed inside the index.html file.

Objective: Build an app that is functionally similar to https://pokemon-search-app.freecodecamp.rocks.